# src/main/java/com/example/orderassignment/dto/AllOrdersDTO.java
package com.example.orderassignment.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
public class AllOrdersDTO {

    private Long orderID;
    @NotNull(message = "Email cannot be null")
    @NotBlank(message = "Email cannot be blank")
    @Email
    private String email;
    private String firstName;
    private String lastName;
    private ProductDTO product;

}