# src/main/java/com/example/orderassignment/dto/ProductDTO.java
package com.example.orderassignment.dto;

import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
public class ProductDTO {

    private Long productID;

    private String name;
}